Resolution Scaling Fix
Now game add black bars to whenever the resolution is not 4:3
