+++ Assets/Scenes/Levels/Intro Loading Scene.unity
@@ -433,6 +433,7 @@ GameObject:
   - component: {fileID: 650174787}
   - component: {fileID: 650174786}
   - component: {fileID: 650174788}
+  - component: {fileID: 650174789}
   m_Layer: 0
   m_Name: Main Camera
   m_TagString: MainCamera
@@ -519,6 +520,20 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   attenuationObject: {fileID: 0}
+--- !u!114 &650174789
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 650174784}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: f195612dd0cf179468e0a175593b6e91, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  targetRatioX: 4
+  targetRatioY: 3
 --- !u!1001 &766777035
 PrefabInstance:
   m_ObjectHideFlags: 0
