+++ Assets/Scenes/Levels/Weapon Select/Loadout Select.unity
@@ -1815,7 +1815,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 1, y: 0}
   m_AnchorMax: {x: 1, y: 0}
-  m_AnchoredPosition: {x: -0.000061035156, y: 0}
+  m_AnchoredPosition: {x: 0, y: 0}
   m_SizeDelta: {x: 167.90002, y: 85.3}
   m_Pivot: {x: 1, y: 0}
 --- !u!114 &914332079
@@ -2131,8 +2131,8 @@ GameObject:
   m_Component:
   - component: {fileID: 1146201595}
   - component: {fileID: 1146201594}
-  - component: {fileID: 1146201597}
   - component: {fileID: 1146201598}
+  - component: {fileID: 1146201599}
   m_Layer: 0
   m_Name: Main Camera
   m_TagString: MainCamera
@@ -2207,19 +2207,6 @@ Transform:
   - {fileID: 327523054}
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 3.766, y: 0, z: 0}
---- !u!114 &1146201597
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1146201592}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 7ee9754eddc646a4997eee41c9a58cc5, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  targetAspect: 1.3333334
 --- !u!114 &1146201598
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -2233,6 +2220,20 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   attenuationObject: {fileID: 0}
+--- !u!114 &1146201599
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1146201592}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: f195612dd0cf179468e0a175593b6e91, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  targetRatioX: 4
+  targetRatioY: 3
 --- !u!1 &1172783643
 GameObject:
   m_ObjectHideFlags: 0
@@ -3567,9 +3568,9 @@ Canvas:
   m_GameObject: {fileID: 1793519759}
   m_Enabled: 1
   serializedVersion: 3
-  m_RenderMode: 0
-  m_Camera: {fileID: 0}
-  m_PlaneDistance: 100
+  m_RenderMode: 1
+  m_Camera: {fileID: 1146201594}
+  m_PlaneDistance: 1
   m_PixelPerfect: 0
   m_ReceivesEvents: 1
   m_OverrideSorting: 0
